requirements.txt
attrs==25.3.0
black==25.1.0
blinker==1.9.0
cattrs==25.1.1
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
filelock==3.18.0
Flask==3.1.1
fsspec==2025.5.1
idna==3.10
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
llvmlite==0.44.0
lsprotocol==2023.0.1
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
mypy_extensions==1.1.0
networkx==3.5
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
openai-whisper @ git+https://github.com/openai/whisper.git@c0d2f624c09dc18e709e37c2ad90c039a4eb72a2
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pygls==1.3.1
regex==2024.11.6
requests==2.32.4
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
tiktoken==0.9.0
torch==2.7.1
tqdm==4.67.1
triton==3.3.1
typing_extensions==4.14.0
urllib3==2.5.0
Werkzeug==3.1.3
